requirements.txt
streamlit
torch
faiss-cpu
numpy
pandas
transformers
sentence-transformers
